fix: ensure directory exists

+++ src/index.ts
@@ -1,17 +1,32 @@
 #!/usr/bin/env node
 
-import { unlink } from 'fs/promises'
-import { join } from 'path'
+import { mkdir, stat, unlink } from 'node:fs/promises'
+import { join } from 'node:path'
 import files from '@unvented/files'
 import minimist from 'minimist'
 
 (async () => {
   const argv = minimist(process.argv.slice(2))
 
+  const stats = async (path: string) => {
+    try {
+      const s = await stat(path)
+  
+      return s
+    } catch (err) {
+      return undefined
+    }
+  }
+
   for (let directory of argv._) {
     directory = join(process.cwd(), directory.startsWith('./') ? directory : `./${directory}`)
 
-    for await (const file of files(directory))
-      await unlink(file)
+    if (!(await stats(directory))?.isDirectory())
+      await mkdir(directory, {
+        recursive: true
+      })
+    else
+      for await (const file of files(directory))
+        await unlink(file)
   }
 })()
